Add basic infrastructure to generate heat map.
Each relation has a heat map, which is actually MAP_SIZE size array of
char. Each MAP_SIZE_RANGE blocks belong to the same range. The current
code report the status of a heatmap whenever increment its temprature.

The most challenge I have is to map relfilenode to OID. It seems the
similar component to snapshot builder will be required.